+++ astropy/io/ascii/tests/test_ecsv.py
@@ -458,7 +458,7 @@ def test_round_trip_masked_table_default(tmpdir):
     t.write(filename)
 
     t2 = Table.read(filename)
-    assert t2.masked is True
+    assert t2.masked is False
     assert t2.colnames == t.colnames
     for name in t2.colnames:
         # From formal perspective the round-trip columns are the "same"
@@ -487,7 +487,7 @@ def test_round_trip_masked_table_serialize_mask(tmpdir):
     t.write(filename, serialize_method='data_mask')
 
     t2 = Table.read(filename)
-    assert t2.masked is True
+    assert t2.masked is False
     assert t2.colnames == t.colnames
     for name in t2.colnames:
         assert np.all(t2[name].mask == t[name].mask)